+++ lib/CountDown/styled.js
@@ -7,7 +7,7 @@ import {
 } from "../styles/colors";
 import { p_12_normal, p_14_medium, p_14_normal } from "../styles/textStyles";
 
-export const CircleWrapper = styled.div`
+export const CountDownWrapper = styled.div`
   display: inline-flex;
 `;
